moved graphs into classes, began implementing sorting
I moved the drawing of graphs from functions into classes, this should give flexibility for sorting.
I stopped in the middle of trying to make the sort button into a toggle for the data, that would show an X mark and use sort(array) to rearrange the data, but didn't finish. Look for boolean variables to continue working.